update demo about webSocket
