+++ sping-boot-websocket-demo/src/main/resources/templates/websocket.html
@@ -30,7 +30,7 @@ websocket测试<br>
     // 判断当前浏览榕是否支持 WebSocket
     if ('WebSocket' in window) {
         // 创建 WebSocket 对象，连接服务器端点
-        websocket = new WebSocket("ws://localhost:8080/ws");
+        websocket = new WebSocket("ws://localhost:8989/ws");
     } else {
         alert('Not support websocket')
     }
